feat(calendar): enhance calendar context with localized date labels
- Removed deprecated 'style' prop from ExcludedProps in calendar types.
- Added localization support for month, month range, and year labels in CalendarCustomContext.
- Refactored CalendarHeader to utilize new text slots for improved accessibility and maintainability.
